==> R/02_load.qmd <==
---
format: 
  html:
    self-contained: true
---

# Create a copy of the raw dataset for analysis

```{r}
library("tidyverse")
library("here")
```


Having the dataset we now verify that it can be loaded without problems, after doing so we create a copy of it in the data folder (the \_raw should be left alone not touched by anyone).

```{r}
# verify that the dataset can be loaded successfully
data_path <- "../_raw/abundance.csv"
dataset <- read_csv(data_path, show_col_types = FALSE)
```

We immediately get parsing errors...

(NOTE: this was really inconsistent and refreshing the page or restarting the R session was sometimes enough to change the behavior).

```{r}
problems(dataset)
```

We can see that they are due to some cells having the a different value type from the expected. We will firstly perform data selection and cleaning, if the issues persist we will handle them later.

```{r}
# show a couple of rows from the dataset just to be sure
head(dataset)
```

```{r}
target_dir <- "../data"
# save the dataset in the data fodler
write_csv(dataset, str_c(target_dir, "/02_dat_load.csv"))
```
