==> R/06_augment.qmd <==
---
format: 
  html:
    self-contained: true
---
# Augmentation

```{r}
library("tidyverse")
library("here")
```


```{r}
data_path <- here("data/05_dat_tidy.RData")
load(data_path)
```

Create a variable which groups together samples from patients having type 2 diabetes or being in the pre-diabetes state.

```{r}
dataset <- dataset |> 
  mutate(is_diseased = ifelse(glucose_tolerance == "t2d" |
                                glucose_tolerance == "impaired",
                              1,
                              0)) |>
  relocate(is_diseased,
           .after = glucose_tolerance)
```

Create a categorical classification based on the BMI.

```{r}
dataset <- dataset |> 
  mutate(BMI_class = case_when(
    bmi < 18.5 ~ "Underweight",
    bmi >= 18.5 & bmi < 25 ~ "Normal weight",
    bmi >= 25 & bmi < 30 ~ "Overweight",
    bmi >= 30 ~ "Obese"
  ) |> 
  factor(levels = c("Underweight", "Normal weight", "Overweight", "Obese"))) |>
  relocate(BMI_class,
           .after = bmi)
```

Create a variable to hold the ratio between high-density and low-density lipoproteins. This ratio is studied and used as a predictor for the insurgence of many cardiovascular diseases, the desired ratio is \>1 as hdl are the good fats while ldl are the bad ones.

```{r}
dataset <- dataset |> 
  mutate(hdl_to_ldl_ratio = hdl / ldl) |> 
  relocate(hdl_to_ldl_ratio,
           .after = ldl)
```

Create a variable for the HOMA-IR score (HOMA-IR (Homeostatic model assessment for insulin resistance) and its categories, a measure insulin resistance / sensitivity ([reference](https://doi.org/10.3390/medicina55050158)).

```{r}
dataset <- dataset |>
  mutate(`HOMA-IR` = fasting_glucose * fasting_insulin / 22.5,
         HOMA_category = case_when(
           `HOMA-IR` < 1.5 ~ "Healthy",
           `HOMA-IR` >= 1.5 & `HOMA-IR` < 2.5 ~ "At risk",
           `HOMA-IR` >= 2.5 & `HOMA-IR` < 5.0 ~ "Insulin Resistant",
           `HOMA-IR` >= 5.0 ~ "Severely Insulin Resistant"
         )) |>
  mutate(HOMA_category = factor(HOMA_category,
                                levels = c("Healthy", "At risk", "Insulin Resistant", "Severely Insulin Resistant"))) |> 
  relocate(`HOMA-IR`,
           .after = fasting_insulin) |> 
  relocate(HOMA_category,
           .after = `HOMA-IR`)
```

Another adjustment to do regards the abundance data, as it right now is composite which does not allow for direct comparison across samples. To prepare for future analysis we apply a center log ratio transformation.

```{r}
metadata <- dataset |> 
  select(-starts_with("s_"))
abundances <- dataset |> 
  select(starts_with("s_"))

rm(dataset)
```

```{r}
# verify that the rows sum to 1
abundances |> 
  rowwise() |> 
  mutate(row_sum = sum(c_across(everything()))) |> 
  select(row_sum)

# apply CLR
CLR_tranformed <- abundances |> 
  mutate(across(starts_with("s_"), ~ ifelse(. == 0, 1e-6, .))) |> 
  rowwise() |> 
  mutate(geom_mean = exp(mean(log(c_across(starts_with("s_")))))) |> 
  mutate(across(starts_with("s_"),~ log(. / geom_mean))) |> 
  select(-geom_mean)
```

```{r}
dataset <- metadata |> 
  bind_cols(CLR_tranformed)
```

```{r}
target_dir <- "../data"

write_csv(dataset, str_c(target_dir,
                         "/06_dat_augmented.csv"))

file_path <- str_c(target_dir, "/06_dat_augmented.RData")
save(dataset, file = file_path)
```
